NODE-80: move validateBlockCheckpoint method from BlockGenerator to ExecEngineUtil since it is also used in Casper.scala, and bring back comment code.
