+++ casper/src/test/scala/io/casperlabs/casper/ValidateTest.scala
@@ -588,7 +588,7 @@ class ValidateTest
         _   <- casperSmartContractsApi.setBonds(bonds)
         dag <- blockDagStorage.getRepresentation
         // FIXME: we should insert the TransformEntry into blockStore, now we simply return empty TransformEntry, this is not correct
-        _ <- BlockGenerator
+        _ <- ExecEngineUtil
               .validateBlockCheckpoint[Task](
                 genesis,
                 dag
