+++ casper/src/test/scala/io/casperlabs/casper/helper/BlockGenerator.scala
@@ -99,87 +99,6 @@ object BlockGenerator {
                )
     } yield result
 
-  //Returns (None, checkpoints) if the block's tuplespace hash
-  //does not match the computed hash based on the deploys
-  def validateBlockCheckpoint[F[_]: Sync: Log: BlockStore: ExecutionEngineService](
-      b: BlockMessage,
-      dag: BlockDagRepresentation[F]
-  ): F[Either[BlockException, Option[StateHash]]] = {
-    val preStateHash = ProtoUtil.preStateHash(b)
-    val tsHash       = ProtoUtil.tuplespace(b)
-    val deploys      = ProtoUtil.deploys(b).flatMap(_.deploy)
-    val timestamp    = Some(b.header.get.timestamp) // TODO: Ensure header exists through type
-    for {
-      parents                              <- ProtoUtil.unsafeGetParents[F](b)
-      processedHash                        <- processDeploys(parents, dag, deploys)
-      (computePreStateHash, deployResults) = processedHash
-      _                                    <- Log[F].info(s"Computed parents post state for ${PrettyPrinter.buildString(b)}.")
-      result <- processPossiblePreStateHash[F](
-                 preStateHash,
-                 tsHash,
-                 deployResults,
-                 computePreStateHash,
-                 timestamp,
-                 deploys
-               )
-    } yield result
-  }
-
-  private def processPossiblePreStateHash[F[_]: Sync: Log: BlockStore: ExecutionEngineService](
-      preStateHash: StateHash,
-      tsHash: Option[StateHash],
-      deployResults: Seq[DeployResult],
-      computedPreStateHash: StateHash,
-      time: Option[Long],
-      deploys: Seq[DeployData]
-  ): F[Either[BlockException, Option[StateHash]]] =
-    if (preStateHash == computedPreStateHash) {
-      processPreStateHash[F](
-        preStateHash,
-        tsHash,
-        deployResults,
-        computedPreStateHash,
-        time,
-        deploys
-      )
-    } else {
-      Log[F].warn(
-        s"Computed pre-state hash ${PrettyPrinter.buildString(computedPreStateHash)} does not equal block's pre-state hash ${PrettyPrinter
-          .buildString(preStateHash)}"
-      ) *> Right(none[StateHash]).leftCast[BlockException].pure[F]
-    }
-
-  private def processPreStateHash[F[_]: Sync: Log: BlockStore: ExecutionEngineService](
-      preStateHash: StateHash,
-      tsHash: Option[StateHash],
-      processedDeploys: Seq[DeployResult],
-      possiblePreStateHash: StateHash,
-      time: Option[Long],
-      deploys: Seq[DeployData]
-  ): F[Either[BlockException, Option[StateHash]]] = {
-    val commutingEffects = findCommutingEffects(processedDeploys)
-    val transforms       = commutingEffects.unzip._1.flatMap(_.transformMap)
-    ExecutionEngineService[F].commit(preStateHash, transforms).flatMap {
-      case Left(ex) =>
-        Log[F].warn(s"Found unknown failure") *> Right(none[StateHash])
-          .leftCast[BlockException]
-          .pure[F]
-      case Right(computedStateHash) =>
-        if (tsHash.contains(computedStateHash)) {
-          //state hash in block matches computed hash!
-          Right(Option(computedStateHash))
-            .leftCast[BlockException]
-            .pure[F]
-        } else {
-          // state hash in block does not match computed hash -- invalid!
-          // return no state hash, do not update the state hash set
-          Log[F].warn(
-            s"Tuplespace hash ${PrettyPrinter.buildString(tsHash.getOrElse(ByteString.EMPTY))} does not match computed hash ${PrettyPrinter
-              .buildString(computedStateHash)}."
-          ) *> Right(none[StateHash]).leftCast[BlockException].pure[F]
-        }
-    }
-  }
 }
 
 trait BlockGenerator {
